Update README.md
It is enough for the program to show a sequence of actions, it will record them and reproduce them at different speeds. Using the Hetic X.1000 computer program, you can create your own autoclicker that repeats the actions of a person on a computer or a bot for a game.